Add files via upload
Project Summary: The concept behind this project is to create a library management system that is capable to issue books and let consumers check different books and their titles categorically. It keeps track of all the details about the books in the library, their price, status, and the total number of books available in the Library. The user will find this automated system easy instead of using the manual writing system.